requirements.txt
Django~=4.2.9
psycopg2==2.9.9
requests==2.31.0
djangorestframework~=3.14.0
celery==5.3.6
rq~=1.15.1
ujson>=4.1.0
collection==0.1.6
djangorestframework-simplejwt==5.3.1
django-webpack-loader==3.0.1
social-auth-app-django==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
redis==4.3.4
kombu~=5.3.5
django-webpush==0.3.5
django-push-notifications~=3.0.2
drf-yasg~=1.21.7
django-environ==0.11.2